add default and material magic strategies
