+++ projects/reactive-form-errors/src/lib/tokens/error-component.token.ts
@@ -1,7 +1,7 @@
 import { InjectionToken } from '@angular/core';
 import { DefaultErrorComponent } from '../components/default-error.component';
 
-export const ERROR_COMPONENT = new InjectionToken('FORM ERROR COMPONENT', {
+export const MAGIC_ERROR_COMPONENT = new InjectionToken('MAGIC FORM ERROR COMPONENT', {
   providedIn: 'root',
   factory: () => DefaultErrorComponent
 });
