+++ projects/reactive-form-errors/src/lib/tokens/index.ts
@@ -1,2 +1,2 @@
-export { FORM_DEFAULT_ERRORS } from './default-errors.token';
-export { ERROR_COMPONENT } from './error-component.token';
+export { FORM_MAGIC_ERRORS } from './default-errors.token';
+export { MAGIC_ERROR_COMPONENT } from './error-component.token';
