+++ projects/reactive-form-errors/src/public_api.ts
@@ -2,7 +2,8 @@
  * Public API Surface of reactive-form-errors
  */
 
-export * from './lib/reactive-form-errors.module';
+export * from './lib/magic-errors.module';
 export * from './lib/components/control-error.interface';
-export { FORM_DEFAULT_ERRORS, ERROR_COMPONENT } from './lib/tokens';
+export { FORM_MAGIC_ERRORS, MAGIC_ERROR_COMPONENT } from './lib/tokens';
 export { FormSubmit } from './lib/directives/form-submit.directive';
+export * from './lib/services/magic-strategy.service';
